add reviews to albums and ll, add main page for current album

+++ lib/listen_lists/reviews/review.ex
@@ -7,7 +7,7 @@ defmodule ListenLists.Reviews.Review do
 
   schema "reviews" do
     field :comment, :string
-    field :rating, :integer
+    field :rating, :float
     belongs_to :user, User
     belongs_to :album, Album
